+++ main/config.js
@@ -1,0 +1,7 @@
+var path = require('path')
+
+module.exports = {
+  APP_NAME: 'WebTorrent',
+  INDEX: 'file://' + path.resolve(__dirname, '..', 'renderer', 'index.html'),
+  startTime: Date.now()
+}
