+++ renderer/lib/capture-video-frame.js
