+++ renderer/lib/torrent-poster.js
