+++ renderer/views/app.js
