+++ renderer/views/player.js
